+++ experimental/unitsconversion/src/measureunit.rs
@@ -13,14 +13,21 @@ use crate::{
     ConversionError,
 };
 
-// TODO(#4369): split this struct to two structs: MeasureUnitParser for parsing the identifier and MeasureUnit to represent the unit.
-// TODO NOTE: the MeasureUnitParser takes the trie and the ConverterFactory takes the full payload and an instance of MeasureUnitParser.
-pub struct MeasureUnit {
-    /// Contains the processed units.
-    pub contained_units: SmallVec<[MeasureUnitItem; 8]>,
+// TODO: add test cases for this parser after adding UnitsTest.txt to the test data.
+/// A parser for the CLDR unit identifier (e.g. `meter-per-square-second`)
+pub struct MeasureUnitParser<'data> {
+    /// Contains the payload.
+    payload: &'data ZeroTrie<ZeroVec<'data, u8>>,
 }
 
-impl MeasureUnit {
+impl<'data> MeasureUnitParser<'data> {
+    // TODO: revisit the public nature of the API. Maybe we should make it private and add a function to create it from a ConverterFactory.
+    /// Creates a new MeasureUnitParser from a ZeroTrie payload.
+    #[cfg(feature = "datagen")]
+    pub fn from_payload(payload: &'data ZeroTrie<ZeroVec<'data, u8>>) -> Self {
+        Self { payload }
+    }
+
     // TODO: complete all the cases for the prefixes.
     // TODO: consider using a trie for the prefixes.
     /// Extracts the SI prefix.
@@ -66,12 +73,11 @@ impl MeasureUnit {
     fn get_unit_id<'data>(
         part: &'data str,
         identifier_split: &mut std::str::Split<'data, char>,
-        trie: &ZeroTrie<ZeroVec<'data, u8>>,
     ) -> Option<usize> {
         let mut part = part.to_string();
 
         loop {
-            if let Some(unit_id) = trie.get(&part) {
+            if let Some(unit_id) = self.payload.get(&part) {
                 return Some(unit_id);
             }
 
@@ -91,10 +97,10 @@ impl MeasureUnit {
     /// For example, if the whole identifier is: "square-kilometer-per-second",
     /// this function will be called for "square-kilometer" with sign (1) and "second" with sign (-1).
     fn analyze_identifier_part(
+        &self,
         identifier_part: &str,
         sign: i8,
         result: &mut Vec<MeasureUnitItem>,
-        trie: &ZeroTrie<ZeroVec<'_, u8>>,
     ) -> Result<(), ConversionError> {
         if identifier_part.is_empty() {
             return Ok(());
@@ -112,7 +118,8 @@ impl MeasureUnit {
             };
 
             let (si_prefix, identifier_after_si) = Self::get_si_prefix(part);
-            let unit_id = Self::get_unit_id(identifier_after_si, &mut identifier_split, trie)
+            let unit_id = self
+                .get_unit_id(identifier_after_si)
                 .ok_or(ConversionError::InvalidUnit)?;
 
             result.push(MeasureUnitItem {
@@ -127,9 +134,9 @@ impl MeasureUnit {
 
     // TODO: add test cases for this function.
     /// Process an identifier.
-    pub fn try_from_identifier<'data>(
+    pub fn try_from_identifier(
+        &self,
         identifier: &'data str,
-        trie: &ZeroTrie<ZeroVec<'data, u8>>,
     ) -> Result<Vec<MeasureUnitItem>, ConversionError> {
         if identifier.starts_with('-') {
             return Err(ConversionError::InvalidUnit);
@@ -142,8 +149,14 @@ impl MeasureUnit {
 
         let mut measure_unit_items = Vec::<MeasureUnitItem>::new();
 
-        Self::analyze_identifier_part(num_part, 1, &mut measure_unit_items, trie)?;
-        Self::analyze_identifier_part(den_part, -1, &mut measure_unit_items, trie)?;
+        self.analyze_identifier_part(num_part, 1, &mut measure_unit_items)?;
+        self.analyze_identifier_part(den_part, -1, &mut measure_unit_items)?;
         Ok(measure_unit_items)
     }
+}
+
+// TODO NOTE: the MeasureUnitParser takes the trie and the ConverterFactory takes the full payload and an instance of MeasureUnitParser.
+pub struct MeasureUnit {
+    /// Contains the processed units.
+    pub contained_units: SmallVec<[MeasureUnitItem; 8]>,
 }